email-gadget: Persistent data is now an overlay
I was trying really hard to keep the data partition as a separate
mountpoint, however, it's pretty much impossible to handle locally
administered users with this scheme. Finally make the data partition now
contain the home folders as a btrfs subvolume, and two other subvolumes
that are mounted as writable overlays on top of /etc and /var at boot
time.