gadget: Patch compilation error in busybox ncurses check
There's a compilation test in busybox that ensures ncurses is available
by trying to compile a simple, empty C file against the library.
Unfortunately, there's a really basic syntax error in this test that is
no longer accepted by GCC, which prevents the do_menuconfig task from
working. Add a patch that fixes the bug.