gadget: Add cloud-mutable-gadget
This a/b mutable gadget is compatible with Linode cloud instances. It
boots using a GRUB2 configuration, and each rootfs partition is a
separate disk. Contrary to popular wisdom, a "disk" in Linode is a
container for a raw partition image, not a disk image (no partition
table). In addition to the gadget image, the cloud-boot-image is also
generated (to be written into the boot "disk"). Ext4.gz images are
generated, which can be uploaded as custom images in the Linode
management portal.