cloud-mutable-gadget: qemu-mutable-gadget: Search for rootfs
Previously, we used hard-coded disk names in the RAUC and GRUB
configuration. This presents problems in ACPI-based systems with
multiple SCSI-attached disks, because the enumeration order of these
disks is not fixed and their names are not stable. Instead, use
partition labels to identify the system0 and system1 slots.
Unfortunately, this also requires adding a RAUC handler that keeps the
slot partition name updated after a software update on these systems.