email-gadget: Improve postfix configuration customization
Previously, the postfix main.cf configuration file was an m4 template.
This was not flexible, in that it didn't allow deployments to change
arbitrary configuration. Now, there is a main.cf.base (installed by us)
and a main.cf.local which is simply appended to the base to form
main.cf. This file can contain arbitrary (valid) postfix configuration
that overrides the configuration parameters in the base file.