gadget: Use an initramfs for all gadgets
I need the root subvolume of the gadget-data partition to be mounted at
/data, in order to integrate gadgets with my backup server. Since the
preinit script generated by the overlay-etc IMAGE_FEATURE does not offer
this capability, extend the existing initramfs script used by the cloud
gadget image to also mount the gadget-data partition. This requires a
migration of the data on the @etc subvolume in order to upgrade.